+++ CHANGELOG.md
@@ -1,10 +1,19 @@
+### v0.4.1:
+#### date: 2014-04-02
+##### changes:
+* Fixed issue [#19](https://github.com/larsonjj/generator-yeogurt/issues/19)
+* Updated file structure so that all dashboard files stay within the `dashboard/` folder (was confusing before when dashboard files would be at the root level)
+* Added [grunt-uncss](https://github.com/addyosmani/grunt-uncss) task when using Vanilla CSS, so unused CSS will be removed during builds (big boost in performance)
+* Cleaned up prompts a bit. Removed miniscule options (ex. adding IE11 icons or adding apple homescreen options: these are now included in the HTML5 Boilerplate extras option)
+* Added a `grunt serve:dist` task, so now it will be easier to see your builds running on a local server.
+
 ### v0.4.0:
 #### date: 2014-04-01
 ##### changes:
 * Updated Bower and NPM dependencies
 * Added new option to use Vanilla CSS
 * Added new option to use Vanilla JavaScript
-* Moved Author and Vendor JavaScript right before `</body>` (Reasoning: http://developer.yahoo.com/performance/rules.html)
+* Moved Author and Vendor JavaScript right before `</body>` ([Reasoning](http://developer.yahoo.com/performance/rules.html))
 * Updated Script subgenerator to no longer create inline scripts
 * Switched to use official Bootstrap SASS Library
 * Fixed Browserify Livereload issue (browserify script changes didn't trigger a rebuild)